+++ libs/discord/cache.c
@@ -8,7 +8,7 @@ static struct Cache guilds = {0};
 
 void clear_cache(struct Cache *cache) {
 	if (cache != NULL) {
-		for (size_t i = 0; i < cache->size; ++i) {
+		for (unsigned long i = 0; i < cache->size; ++i) {
 			free(cache->data[i]);
 		}
 
@@ -24,9 +24,9 @@ void add_to_cache(struct Cache *cache, const char *data) {
 	strcpy(cache->data[cache->size - 1], data);
 }
 
-void remove_from_cache_index(struct Cache *cache, const size_t index) {
-	for (size_t i = (index + 1); i < cache->size; ++i) {
-		const size_t length = strlen(cache->data[i]);
+void remove_from_cache_index(struct Cache *cache, const unsigned long index) {
+	for (unsigned long i = (index + 1); i < cache->size; ++i) {
+		const unsigned long length = strlen(cache->data[i]);
 		cache->data[i - 1] = allocate(cache->data[i - 1], -1, length + 1, sizeof(char));
 		strcpy(cache->data[i - 1], cache->data[i]);
 	}
