+++ libs/json/set.c
@@ -15,7 +15,7 @@ jsonelement_t *create_empty_json_element(const bool is_array) {
 
 static void set_value(jsonelement_t *element, void *value, const unsigned char type) {
 	if (element->type == JSON_ARRAY || element->type == JSON_OBJECT) {
-		for (size_t i = 0; i < element->size; ++i) {
+		for (unsigned long i = 0; i < element->size; ++i) {
 			json_free(((jsonelement_t **) element->value)[i], false);
 		}
 	}
@@ -23,7 +23,7 @@ static void set_value(jsonelement_t *element, void *value, const unsigned char t
 	element->type = type;
 
 	if (type == JSON_STRING) {
-		const size_t value_length = strlen(value);
+		const unsigned long value_length = strlen(value);
 
 		element->value = allocate(element->value, -1, value_length + 1, sizeof(char));
 		strcpy(element->value, value);
@@ -38,7 +38,7 @@ static void set_value(jsonelement_t *element, void *value, const unsigned char t
 		element->size = array->size;
 		element->value = allocate(element->value, -1, array->size, sizeof(jsonelement_t));
 
-		for (size_t i = 0; i < array->size; ++i) {
+		for (unsigned long i = 0; i < array->size; ++i) {
 			((jsonelement_t **) element->value)[i] = allocate(NULL, -1, 1, sizeof(jsonelement_t));
 			jsonelement_t *array_element = ((jsonelement_t **) array->value)[i];
 			jsonelement_t *sub_element = ((jsonelement_t **) element->value)[i];
@@ -56,7 +56,7 @@ static void set_value(jsonelement_t *element, void *value, const unsigned char t
 		element->size = object->size;
 		element->value = allocate(element->value, -1, object->size, sizeof(jsonelement_t *));
 
-		for (size_t i = 0; i < object->size; ++i) {
+		for (unsigned long i = 0; i < object->size; ++i) {
 			((jsonelement_t **) element->value)[i] = allocate(NULL, -1, 1, sizeof(jsonelement_t));
 			jsonelement_t *object_element = ((jsonelement_t **) object->value)[i];
 			jsonelement_t *sub_element = ((jsonelement_t **) element->value)[i];
@@ -86,12 +86,12 @@ void json_set_val(jsonelement_t *target, const char *key, void *value, const cha
 		element = ((jsonelement_t **) element->value)[element->size - 1];
 		set_value(element, value, type);
 	} else {
-		const size_t key_length = strlen(key);
+		const unsigned long key_length = strlen(key);
 		struct Split splitter = split(key, key_length, ".");
 
-		for (size_t i = 0; i < splitter.size; ++i) {
+		for (unsigned long i = 0; i < splitter.size; ++i) {
 			const char *skey = splitter.data[i].data;
-			const size_t skey_length = splitter.data[i].length;
+			const unsigned long skey_length = splitter.data[i].length;
 
 			char bwp[skey_length - 2 + 1];
 			bwp[skey_length - 2] = 0;
@@ -101,7 +101,7 @@ void json_set_val(jsonelement_t *target, const char *key, void *value, const cha
 			if (skey[0] == '[' && skey[skey_length - 1] == ']' && (iskey > 0 || strcmp(bwp, "0") == 0)) {
 				if (element->type != JSON_ARRAY) {
 					if (element->type == JSON_OBJECT) {
-						for (size_t n = 0; n < element->size; ++n) {
+						for (unsigned long n = 0; n < element->size; ++n) {
 							json_free(((jsonelement_t **) element->value)[n], false);
 						}
 					}
@@ -124,7 +124,7 @@ void json_set_val(jsonelement_t *target, const char *key, void *value, const cha
 					char new_key[key_length - skey_length + 1];
 					struct Join new_key_joins[splitter.size - 1];
 
-					for (size_t n = 1; n < splitter.size; ++n) {
+					for (unsigned long n = 1; n < splitter.size; ++n) {
 						new_key_joins[n - 1].data = splitter.data[n].data;
 						new_key_joins[n - 1].length = splitter.data[n].length;
 					}
@@ -137,7 +137,7 @@ void json_set_val(jsonelement_t *target, const char *key, void *value, const cha
 			} else {
 				if (element->type != JSON_OBJECT) {
 					if (element->type == JSON_ARRAY) {
-						for (size_t n = 0; n < element->size; ++n) {
+						for (unsigned long n = 0; n < element->size; ++n) {
 							json_free(((jsonelement_t **) element->value)[n], false);
 						}
 					}
@@ -148,7 +148,7 @@ void json_set_val(jsonelement_t *target, const char *key, void *value, const cha
 
 				long at = -1;
 
-				for (size_t n = 0; n < element->size; ++n) {
+				for (unsigned long n = 0; n < element->size; ++n) {
 					if (strcmp(((jsonelement_t **) element->value)[n]->key, skey) == 0) {
 						at = n;
 						n = element->size;
@@ -175,7 +175,7 @@ void json_set_val(jsonelement_t *target, const char *key, void *value, const cha
 					char new_key[key_length - skey_length + 1];
 					struct Join new_key_joins[splitter.size - 1];
 
-					for (size_t n = 1; n < splitter.size; ++n) {
+					for (unsigned long n = 1; n < splitter.size; ++n) {
 						new_key_joins[n - 1].data = splitter.data[n].data;
 						new_key_joins[n - 1].length = splitter.data[n].length;
 					}
