+++ libs/utils/join.c
@@ -3,11 +3,11 @@
 
 #include <utils.h>
 
-size_t calculate_join(const struct Join *value, unsigned short size, const char *separator) {
-	size_t source_length = (strlen(separator) * (size - 1));
+unsigned long calculate_join(const struct Join *value, unsigned short size, const char *separator) {
+	unsigned long source_length = (strlen(separator) * (size - 1));
 
 	for (unsigned short i = 0; size != 0; ++i) {
-		const size_t string_length = value[i].length;
+		const unsigned long string_length = value[i].length;
 
 		if (string_length != 0) {
 			source_length += string_length;
@@ -18,9 +18,9 @@ size_t calculate_join(const struct Join *value, unsigned short size, const char 
 	return source_length;
 }
 
-size_t join(const struct Join *value, char *source, unsigned short size, const char *separator) {
-	const size_t separator_length = strlen(separator);
-	size_t source_length = 0;
+unsigned long join(const struct Join *value, char *source, unsigned short size, const char *separator) {
+	const unsigned long separator_length = strlen(separator);
+	unsigned long source_length = 0;
 
 	for (unsigned short i = 0; size != 0; ++i) {
 		const struct Join join_element = value[i];
@@ -43,8 +43,8 @@ size_t join(const struct Join *value, char *source, unsigned short size, const c
 	return source_length;
 }
 
-void create_join_elements_nz(struct Join *joins, const char **list, const size_t list_size) {
-	for (size_t i = 0; i < list_size; ++i) {
+void create_join_elements_nz(struct Join *joins, const char **list, const unsigned long list_size) {
+	for (unsigned long i = 0; i < list_size; ++i) {
 		joins[i].data = (char *) list[i];
 		joins[i].length = ((list[i] == NULL) ? 0 : strlen(list[i]));
 	}
